+++ src/components/TextField/TextField.js
@@ -39,6 +39,7 @@ const TextField = React.forwardRef((props, ref) => {
     trailingIconClassName,
     label,
     style,
+    menuProps,
     ...otherProps
   } = props
 
@@ -153,6 +154,7 @@ const TextField = React.forwardRef((props, ref) => {
         isOpen={!!anchorElement}
         onClose={handleMenuClose}
         className={classNames(styles.textFieldMenu, menuClassName)}
+        {...menuProps}
       >
         {items.map(item =>
           renderItem ? (
@@ -189,7 +191,7 @@ const TextField = React.forwardRef((props, ref) => {
 
   const inputDefaultValue = type === types.options ? undefined : defaultValue
 
-  if (isView && type === 'password') {
+  if (isView && type === types.password) {
     return
   }
 
@@ -244,6 +246,7 @@ TextField.defaultProps = {
   onClick: () => {},
   onChange: () => {},
   items: [],
+  menuProps: {},
 }
 
 TextField.propTypes = {
@@ -299,6 +302,8 @@ TextField.propTypes = {
   label: propTypes.string,
   /** @ignore */
   style: propTypes.string,
+  /** props to be propagated to menu component*/
+  menuProps: propTypes.object,
 }
 
 export default memo(TextField)